added device monitoring files and some audit refactoring
