+++ src/rtgs_lab_tools/audit/__init__.py
@@ -1,1 +1,7 @@
 """Audit module for RTGS Lab Tools."""
+
+from .audit_service import AuditService
+from .report_service import ReportService
+from .cli import audit_cli
+
+__all__ = ["AuditService", "ReportService", "audit_cli"]
