+++ src/rtgs_lab_tools/data_parser/__init__.py
@@ -1,1 +1,5 @@
 """Universal JSON packet parser for historical packets based on device JSON schema tool for RTGS Lab Tools."""
+
+from .core import parse_gems_data
+
+__all__ = ['parse_gems_data']
