+++ src/rtgs_lab_tools/device_monitoring/cli.py
@@ -1,0 +1,6 @@
+#sets up the cli for the device monitoring module
+#should be one function that takes multiple arguments, but the function is implmented in core, cli is jsut a wrapper
+#arguemnts:
+#start_date: datetime
+#end date is always current time
+#node_ids: list of node ids
