+++ src/rtgs_lab_tools/device_monitoring/core.py
@@ -1,0 +1,4 @@
+#implementation of the core functionality of device monitoring
+#should organize all the functions from within data_getter, date_formatter, data_analyzer, notification system
+#data_getter->data_fromatter->data_analyzer->notification_system
+#should create the function that the cli.py wraps
