+++ src/rtgs_lab_tools/device_monitoring/data_analyzer.py
@@ -1,0 +1,3 @@
+#should take parsed dataframes of different packet types
+#should implement thresholds depending on the value that is being analyzed
+#should return formatted data for notification and visualization
