+++ src/rtgs_lab_tools/device_monitoring/data_formatter.py
@@ -1,0 +1,2 @@
+#should utilize parse_gems_data to parse specific packets that are needed to analyze
+#shoudl return df
