+++ src/rtgs_lab_tools/device_monitoring/data_getter.py
@@ -1,0 +1,2 @@
+#should uitlize get_raw_data from sensing_data module
+#should return a df
