+++ src/rtgs_lab_tools/device_monitoring/notification_system.py
@@ -1,0 +1,1 @@
+#should send email to email address with results of the operation
